[#72661] Singularize Backlogs::BurndownChartController
Renames the namespaced controller, view directory, and spec files to
use the singular `burndown_chart` form, matching the singular URL
segment (`sprints/:sprint_id/burndown_chart`) and the sibling
`Backlogs::TaskboardController`.

The named route helper `project_backlogs_sprint_burndown_chart_path`
is unchanged, so no caller updates are required.

https://community.openproject.org/wp/72661
